+++ src/EntityFrameworkCore/JK.Common.EntityFrameworkCore/Auditing/AuditableEntity.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace JK.Common.EntityFrameworkCore.Auditing
+{
+    public class AuditableEntity
+    {
+        public string CreatedBy { get; set; }
+        public string ModifiedBy { get; set; }
+        public DateTimeOffset DateCreated { get; set; }
+        public DateTimeOffset DateModified { get; set; }
+    }
+}